+++ env.py
@@ -1,5 +1,9 @@
-HOST="localhost";
-PORT="5432";
-STANDARD_DB="database";
-USER="postgres";
-PASSWORD="admin";
+# Backend variables
+DEBUG = True
+
+# Database variables
+HOST = "localhost"
+PORT = "5432"
+STANDARD_DB = "database"
+USER = "postgres"
+PASSWORD = "admin"
