+++ run.py
@@ -1,8 +1,16 @@
-from game.game import *
 from website.flask_website import *
+import cv2
 
-newgame = Game(None)
-newgame.run()
 
-#website = Website()
-#website.run()
+game = Game(None)
+# feed = cv2.VideoCapture('data/video/best yet.mp4')
+feed = cv2.VideoCapture(0, cv2.CAP_DSHOW)
+print(feed)
+print(feed.read())
+# while True:
+#     nextFrame, frame = feed.read()
+#     game.showFrame(frame)
+# game.superRun(feed, local=True)
+game.run(feed)
+game.maarZouDezeWelWerkenDan()
+# Website().run(0)
