--- run_web.py
@@ -1,5 +1,0 @@
-from website.flask_website import *
-
-website = Website()
-
-website.run_video()
